ias-shell: Skip rendering of a surface for local SoC (v2)
This patch addresses a problem: How do we continue to receive
new buffers from a client app even though we aren't going to
display them on the local SoC because they are meant for another
SoC?
If a surface is not meant to be displayed on the local SoC,
then we skip its rendering. We essentially do this by assigning the
role_name field within a surface to be "remote_soc". This can be set
by ias_hmi upon call from an ias_hmi client. It will be checked by
the backend when assigning a view to a plane and we will skip any
client whose role_name field is set to remote_soc, thereby assigning
them to the NULL plane and the renderer will skip drawing their views.

Had we simply set the client's status to hidden, then it would be
removed from the view_list by the compositor and it wouldn't receive
any frame done callbacks which essentially would result in the client
not drawing any more frames.

v2: Made the changes such that we don't need to change the
weston_surface data structure.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Reviewed-by: Mateusz Polrola <mateuszx.potrola@intel.com>