ias: Initial commit
This is the first commit for Intel Automotive Solution Compositor
for Wayland. This compositor showcases how best to use Intel
graphics and display engines for automotive use cases.

A Wayland compositor usually has a shell, a backend and the core
parts (like the renderer). Shell is responsible for the look and
feel. For example, the xdg shell is meant for desktops. We have
replaced it with ias shell for automotive purposes. Backend is
responsible for communicating with the DRM, memory and input
management. We have replaced the default drm-backend with
ias-backend as it show cases how to make use of multiple HW planes,
use display controller specific functionality, offload GPU etc.
Finally, we have introduced a plugin framework which allows a
customer to completely overtake the layout and input management
capabilities of the compositor, such that with a small amount of
code, a layout plugin can overhault the look and feel of the
display from one frame to the next, and with an input plugin, they
can redirect input events to any surface/s.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>