compositor-ias: Only disable HW planes in case we are capturing outputs
Disabling of planes was done whenever we started capturing either a
surface or a Wayland output when it really should only be done for an
output because that's when we need to combine the contents of all of
the surfaces on a single HW plane, encode it and send it to another
SoC.
If we disable planes for a surface as well, then we skip
ias_assign_planes completely which means that we cannot skip local
composition for a surface that may be necessary in orchestration cases.
In compositor.c's, weston_output_repaint function, the code looks like
this:
	if (output->assign_planes && !output->disable_planes) {
		output->assign_planes(output, repaint_data);
	} else {
		wl_list_for_each(ev, &ec->view_list, link) {
			weston_view_move_to_plane(ev, &ec->primary_plane);
			ev->psf_flags = 0;
		}
	}

Due to the fact that we are disabling planes whenever we start encoding
a surface as well, we will always go in the else case which means it will
be locally composited as well and this is precisely what we need to
avoid.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Reviewed-by: Mateusz Polrola <mateuszx.potrola@intel.com>