hdcp: Added polling mechanism to query CP status from the driver
According to the spec, userspace can only request to turn on CP
by setting the flag DRM_MODE_CONTENT_PROTECTION_DESIRED, but it
is up to driver to decide if it really will be enabled and change
that property to DRM_MODE_CONTENT_PROTECTION_ENABLED.
We must check to see if the driver changed this DESIRED to ENABLED
and then report back to our client that CP indeed got turned on.
We do so by immediately checking once if the driver enabled CP. If
not, then we start a timer which will check once every 16 ms for up
to 10 times or earlier if the driver turns on CP by then.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>