RemoteDisplay: Added UDP transport plugin
TCP protocol has an inherent issue that if there is ever a packet
loss, then TCP halves the bandwidth and increments slowly until
full bandwidth as long as it doesn't see any further errors.
However, this is bad for video streaming purposes because it means
that there is a very high chance of losing subsequent frames due to
a loss of bandwidth.
With this patch, we introduce UDP transport plugin for remote display.
UDP works on a fire and forget principle where even if we lose a frame
it has no impact on any subsequent frames. Also, UDP is supposedly
faster than TCP so this may have a positive impact on our latency.
The receiver end needs to be run like this:
gst-launch-1.0 udpsrc port=<PORTID> ! h264parse ! mfxdecode live-mode=true !
mfxsinkelement sync=false

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Reviewed-by: Gowtham Hosamane <gowtham.hosamane@intel.com>