compositor-ias: Fix the issue around FB creation
While capturing a FB for remoting, we were constantly getting
an error from the i915 driver due to invalid capabilities used.
This change gets rid of that error and we can now use
drmModeAddFB2WithModifiers correctly.

NOTE: There is still an issue where the receiver SOC is not
showing the right data yet.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>