RemoteDisplay: UDP streaming using gstreamer pipeline to transport
Moved to using gstreamer pipeline for transportation because our
transport clearly has a problem. While this gstreamer pipeline
approach also has an issue that it increases the CPU load slightly
on the sender side, the good news is that it can work with a gstreamer
receiver. Following AVB plugin's lead.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>