RemoteDisplay: Added native transport and made it an option
Now, we can provide remote-display with an option --tp=native/gst.
This will select the transport mechanism which can either be
gstreamer based or done natively by the UDP transport plugin.

Note that this is an optional parameter. If this option is not
provided, then we default to native transport.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Reviewed-by: Kevron Rees <kevron.m.rees@intel.com>