ias-shell: Print fps now also prints number of flip dones
Previously, print fps would only count and print the number of
commits that an app made. However, it is also important to know
the number of flips that occured for that surface.

Note that this doesn't yet take care of apps that may be on two
different Wayland outputs at the same time.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Reviewed-by: Daniel Charles <daniel.charles@intel.com>