compositor-ias: Disable sprites and cursors only once
There is an issue seen in the DRM core where if we create an
atomic commit with one plane being updated with a new FB and
others like sprites and cursors being turned off, then every
once in a while, the DRM may think that the flip is not done
for the plane being disabled. This is because the DRM considers
the disabled planes to be part of a fake CRTC. It only sends
the flip done for the real CRTC and this is what starts the
repaint cycle for the compositor. So a situation may arise
like this:
1) Compositor created an atomic commit updating FB for plane 1,
disabling plane 2 and cursor.
2) DRM core considers plane 1 as part of CRTC A, but plane 2
as part of fake crtc.
3) DRM sends flip done event back to the compositor for plane 1.
4) DRM doesn't send a flip done event for plane 2.
5) Compositor starts its repaint cycle and provides a new frame
buffer to the DRM for plane 1, again disabling plane 2 and cursor.
6) DRM checks if flip done is finished for all planes. finds that
the fake CRTC's flip done is not finished and returns an EBUSY
back to the compositor.

In the above scenario, the compositor  doesn't necessarily know
how to handle this EBUSY error. It doesn't know if it has to wait
for 1 ms or several Vblanks before which it should try again.
The right fix lies in the DRM. However, we can avoid this problem
from the compositor side as well. Currently, IAS disables any
sprites and cursor planes that it doesn't use for a particular
frame. However, it doesn't need to do it over and over again if
a plane is already disabled once. Just disabling once should be
sufficient (even though there should be nothing wrong with
disabling for every Vblank except for this DRM issue).
So with this change, we only disable unused planes once. If they
are ever used again in the future, we turn on the flag which will
disable these unused planes only once again.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>