wl_base: Added a new API to find out about cp status
A caller who has just set content protection may wish to know
if it was indeed set or not by the compositor. Since the rules
about content protection say that the driver may not be able to
do it instantly, the compositor has to query the driver
periodically to find out if it indeed set it or not. This means
that the compositor may tell wl_base library at a later point
as well. So the caller may also have to poll to find out the
current status.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Reviewed-by: Daniel Charles<daniel.charles@intel.com>